Struct glossy_codegen::Config

pub struct Config { /* fields omitted */ }

Configuration for the glossy shader preprocessor.

Methods

impl Config

fn new(lang: Language) -> Config

Create a new blank configuration for loading GLSL shaders.

fn vertex(self, pattern: &str) -> Config

Specify a glob pattern which adds the matching files to the list of shader sources as vertex shader sources.

fn fragment(self, pattern: &str) -> Config

Specify a glob pattern which adds the matching files to the list of shader sources as fragment shader sources.

fn source(self, pattern: &str) -> Config

Specify a glob pattern which adds the matching files to the list of shader sources. If the extension is "vert" or "frag", they will be loaded as vertex and fragment shaders respectively. If not, they will be added to the list of generic shader files. Thus this function should really only be used to add non-vertex or fragment shader sources, as these are the only kinds that glsl-optimizer can optimize.

fn include(self, pattern: &str) -> Config

Specify a glob pattern which adds the matching files to the list of include files.

fn optimize(self) -> Config

Specify that the shaders should be optimized with glsl-optimizer. This is only available when the 'optimizer' feature is enabled (which is the default option).

This also implicitly sets the discard_line_info() option.

fn discard_line_info(self) -> Config

Do not preserve the line and file information in the shader.

This option, when set, will strip all comments and empty lines, but will not modify whitespace or minify the shader otherwise. Most of the time, optimize() makes more sense than this option, though it might be useful for the very slight minification it can provide for shader sources that aren't supported by glsl-optimizer.

fn allow_untested_versions(self) -> Config

Allow the optimizer to work on untested language versions.

According to the glsl-optimizer readme, versions of GLSL beyond 1.20 are untested. This flag allows glsl-optimizer to try optimizing them regardless.

fn build(self)

Process the specified GLSL shaders.

This function will panic if the source files cannot be written, a recursive inclusion is detected, or if an error occurs during optimization. The optimizer can thus serve as a compile-time shader validator to a certain extent. Note that the error messages generated by the compiler will have inaccurate file line information, as it does not account for the

included files or stripped comments (the optimizer does not support reading comments).